Renamed package to aerospike vector search
Addressed Jesse's review comments.
Fixed index_waiting logic to check index status twice before returning.
Changed index wait to 5 seconds for consistency
