+++ docs/types.rst
@@ -3,7 +3,7 @@ types
 
 This module provides administrative functions for the VectorDB system.
 
-.. automodule:: aerospike_vector.types
+.. automodule:: aerospike_vector_search.types
    :members:
    :undoc-members:
    :show-inheritance:
