+++ docs/vectordb_admin.rst
@@ -4,7 +4,7 @@ vectordb_admin Module
 This module contains the admin client, which is designed to conduct Proximus administrative operation such as creating indexes, querying index information, and dropping indexes.
 
 
-.. automodule:: aerospike_vector.vectordb_admin
+.. automodule:: aerospike_vector_search.vectordb_admin
    :members:
    :undoc-members:
    :show-inheritance:
