+++ docs/vectordb_client.rst
@@ -4,7 +4,7 @@ vectordb_client Module
 This module contains the vector client (VectorDbClient), which specializes in performing database operations with vector data.
 
 
-.. automodule:: aerospike_vector.vectordb_client
+.. automodule:: aerospike_vector_search.vectordb_client
    :members:
    :undoc-members:
    :show-inheritance:
