+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["setuptools","wheel"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "aerospike-vector"
+name = "aerospike-vector-search"
 description = "Aerospike Proximus Client Library for Python"
 authors = [
     { name = "Aerospike, Inc.", email = "info@aerospike.com" }
@@ -22,7 +22,7 @@ classifiers = [
     "Programming Language :: Python :: Implementation :: CPython",
     "Topic :: Database"
 ]
-version = "0.4.0"
+version = "0.5.0dev1"
 requires-python = ">3.8"
 dependencies = [
     "grpcio",
@@ -35,5 +35,5 @@ dependencies = [
 [tool.setuptools]
 zip-safe = false
 include-package-data = true
-packages = ["aerospike_vector"]
+packages = ["aerospike_vector_search", "aerospike_vector_search.aio", "aerospike_vector_search.shared", "aerospike_vector_search.shared.proto_generated", "aerospike_vector_search.internal", "aerospike_vector_search.aio.internal"]
 package-dir={"" = "src"}
