+++ src/aerospike_vector_search/shared/conversions.py
