+++ src/aerospike_vector_search/shared/proto_generated/index_pb2_grpc.py
