+++ src/aerospike_vector_search/shared/proto_generated/transact_pb2_grpc.py
